add test to setup.py+separate tests into separate folder
